Move implementation code for custom item out of header.
[TIMOB-12788]
